requirements.txt
langchain==0.1.12
langchain_community==0.0.28
langchain_core==0.1.31
langchain_openai==0.0.8
langchain_text_splitters==0.0.1
python-dotenv==1.0.1
